requirements.txt
python-binance
TA-lib
numpy
backtrader
flask
